+++ backend/database/migrations/2025_05_19_213020_create_countries_table.php
@@ -15,7 +15,8 @@ return new class extends Migration
     {
         Schema::create('countries', function (Blueprint $table) {
             $table->increments('Countryid');
-            $table->unsignedInteger('Tripid');
+           $table->unsignedBigInteger('Tripid');
+
             $table->string('Name');
             $table->string('Image_path')->nullable();
